+++ src/main/java/client/Game.java
@@ -1,0 +1,10 @@
+package client;
+
+import client.graphics.window.MainWindow;
+
+public class Game extends Thread {
+    public void run() {
+        MainWindow mainWindow = new MainWindow();
+        mainWindow.start();
+    }
+}