+++ changelog.md
@@ -1,5 +1,16 @@
 # Battle City - client side
 
 ## Version 0.1.0
+
 - Project Initialization
 
+## Version 0.2.0
+
+- Implemented:
+  - Movement of the main tank using the arrows or _w, d, s, a_ keys.
+  - _Rotation_ of the main tank according to key that was pressed.
+  - `Start` button to initialize the game. 
+  - Client connection with the server.
+- Pending:
+  - Send tank position to the client.
+  - Receive server approval to perform the movement.